# openapi.yaml
openapi: 3.0.3
info:
  title: ODS API
  description: |-
    Especificación API para manejo de dimensiones, ODS, metas, clases, módulos, profesores, entidades externas e iniciativas.
  termsOfService: http://swagger.io/terms/
  contact:
    email: admin@odsapi.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /iniciativas:
    get:
      tags:
        - iniciativas
      summary: Recuperar todas las iniciativas
      operationId: getIniciativas
      responses:
        '200':
          description: Lista de iniciativas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Iniciativa'
    post:
      tags:
        - iniciativas
      summary: Crear una nueva iniciativa
      operationId: createIniciativa
      requestBody:
        description: Datos de la nueva iniciativa
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIniciativa'
      responses:
        '201':
          description: Iniciativa creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Iniciativa'
  /iniciativas/{idIniciativa}:
    put:
      tags:
        - iniciativas
      summary: Update an existing iniciative
      description: Update an existing activity by Id
      operationId: updateIniciativa
      parameters:
        - name: idIniciativa
          in: path
          description: Activity id to Update
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Update an existent Activity in the GYM
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIniciativa'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Iniciativa'
        '404':
          description: Activity not found
    delete:
      tags:
        - iniciativas
      summary: Deletes a iniciativas
      description: delete a iniciativas
      operationId: deleteIniciativa
      parameters:
        - name: idIniciativa
          in: path
          description: Iniciativa id to Delete
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '404':
          description: Activity not found
components:
  schemas:
    Dimension:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Social
    Ods:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Fin de la Pobreza
        dimension:
            $ref: '#/components/schemas/Dimension'
    Meta:
      type: object
      properties:
        id:
          type: integer
          example: 1
        descripcion:
          type: string
          example: Fomentar la educación digital
        ods:
            $ref: '#/components/schemas/Ods'
    Clase:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Clase de Matemáticas
    Modulo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Álgebra
        clase:
            $ref: '#/components/schemas/Clase'
    Profesor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Juan Pérez
    EntidadExterna:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Universidad Nacional
    Iniciativa:
      type: object
      properties:
        id:
          type: integer
          example: 1
        tipo:
          type: string
          example: Investigación sobre IA
        horas:
          type: integer
          example: 50
        nombre:
          type: string
          example: IA en educación
        explicacion:
          type: string
          example: Informe de IA
        fecha_inicio:
          type: string
          format: date
          example: '2025-01-01'
        fecha_fin:
          type: string
          format: date
          example: '2025-06-01'
        eliminado:
          type: boolean
          example: true
        innovador:
          type: boolean
          example: true
        anyo_lectivo:
          type: string
          example: '2024-2025'
        imagen:
          type: string
          example: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVUGdra5SfyyJUvZ-mWByvoY1gEwp0df8NLA&s'
        fecha_registro:
          type: string
          format: date
          example: '2025-01-01'
        mas_comentarios:
          type: string
          example: Si
        redes_sociales:
          type: array
          items: 
            type: string
          example: linkedin
        metas:
          type: array
          items:
            $ref: '#/components/schemas/Meta'
        profesores:
          type: array
          items:
            $ref: '#/components/schemas/Profesor'
        entidadesExternas:
          type: array
          items:
            $ref: '#/components/schemas/EntidadExterna'
        modulos:
          type: array
          items:
            $ref: '#/components/schemas/Modulo'
    NewIniciativa:
      type: object
      properties:
        tipo:
          type: string
        horas:
          type: integer
        nombre:
          type: string
        explicacion:
          type: string
        fecha_inicio:
          type: string
          format: date
        fecha_fin:
          type: string
          format: date
        eliminado:
          type: boolean
        innovador:
          type: boolean
        anyo_lectivo:
          type: string
        imagen:
          type: string
          format: uri
        fecha_registro:
          type: string
          format: date
          example: '2025-01-01'
        mas_comentarios:
          type: string
        redes_sociales:
          type: array
          items: 
            type: string
        metas:
          type: array
          items:
            type: integer
        profesores:
          type: array
          items:
            type: integer
        entidades_externas:
          type: array
          items:
            type: integer
        modulos:
          type: array
          items:
            type: integer